requirements.txt
fastapi
uvicorn[standard]
pandas
openai
python-dotenv
pydantic
python-multipart
python-docx
xlrd
openpyxl
gunicorn